+++ src/components/ids-data-grid/ids-data-grid.ts
@@ -1,8 +1,7 @@
 // Utils
 import { customElement, scss } from '../../core/ids-decorators';
-import { attributes, IdsDirection } from '../../core/ids-attributes';
+import { attributes } from '../../core/ids-attributes';
 import { stringToBool } from '../../utils/ids-string-utils/ids-string-utils';
-import { next, previous } from '../../utils/ids-dom-utils/ids-dom-utils';
 import { exportToCSV, exportToXLSX } from '../../utils/ids-excel-exporter/ids-excel-exporter';
 import { eventPath, findInPath } from '../../utils/ids-event-path-utils/ids-event-path-utils';
 
@@ -122,6 +121,22 @@ export default class IdsDataGrid extends Base {
     return this.container?.querySelector<HTMLElement>('.ids-data-grid-body');
   }
 
+  /* Returns the outside wrapper element */
+  get wrapper() {
+    return this.container?.parentNode as HTMLElement | undefined | null;
+  }
+
+  /* Returns the cell with active tabindex */
+  get cellFocused(): IdsDataGridCell | null {
+    return this.shadowRoot?.querySelector<IdsDataGridCell>('[tabindex="0"]') ?? null;
+  }
+
+  /* Returns the last active cell recognized by datagrid */
+  get cellLastActive(): IdsDataGridCell | null {
+    const [rowIndex = 0, columnIndex = 0] = this.getAttribute('active-cell')?.split(':') || [];
+    return this.rowByIndex(Number(rowIndex))?.cellByIndex(Number(columnIndex)) ?? null;
+  }
+
   /* Returns all the row elements in an array */
   get rows(): IdsDataGridRow[] {
     // NOTE: Array.from() seems slower than dotdotdot array-destructuring.
@@ -155,11 +170,6 @@ export default class IdsDataGrid extends Base {
    */
   get virtualRows(): Array<Record<string, any>> {
     return this.data.filter((row) => !row.rowHidden);
-  }
-
-  /* Returns the outside wrapper element */
-  get wrapper() {
-    return this.container?.parentNode as HTMLElement | undefined | null;
   }
 
   connectedCallback() {
@@ -260,7 +270,7 @@ export default class IdsDataGrid extends Base {
 
     const html = `<div class="ids-data-grid-wrapper">
         <span class="ids-data-grid-sort-arrows"></span>
-        <div class="ids-data-grid${cssClasses}" role="table" part="table" aria-label="${this.label}" data-row-height="${this.rowHeight}">
+        <div tabindex="0" class="ids-data-grid${cssClasses}" role="table" part="table" aria-label="${this.label}" data-row-height="${this.rowHeight}">
         ${innerTemplate}
         </div>
         ${emptyMesageTemplate}
@@ -758,40 +768,30 @@ export default class IdsDataGrid extends Base {
 
       if (!e.shiftKey) this.#resetLastShiftedRow();
       if (inFilter) return;
-      if (!this.activeCell?.node) return;
       if (this.openMenu) return;
 
-      const cellNode = this.activeCell.node;
-      const cellNumber = Number(this.activeCell?.cell);
-      const rowDiff = key === 'ArrowDown' ? 1 : (key === 'ArrowUp' ? -1 : 0); //eslint-disable-line
-      const cellDiff = key === 'ArrowRight' ? 1 : (key === 'ArrowLeft' ? -1 : 0); //eslint-disable-line
-      const nextRow = Number(next(cellNode.parentElement, `:not([hidden])`)?.getAttribute('row-index'));
-      const prevRow = Number(previous(cellNode.parentElement, `:not([hidden])`)?.getAttribute('row-index'));
-      const rowIndex = key === 'ArrowDown' ? nextRow : prevRow;
+      let nextCell = this.cellLastActive;
+      if (nextCell?.isEditing) return;
 
-      const movingHorizontal = key === 'ArrowLeft' || key === 'ArrowRight';
-      const reachedHorizontalBounds = cellNumber < 0 || cellNumber >= this.visibleColumns.length;
-      if (movingHorizontal && reachedHorizontalBounds) return;
+      e.preventDefault();
+      e.stopPropagation();
 
-      const movingVertical = key === 'ArrowDown' || key === 'ArrowUp';
-      const reachedVerticalBounds = nextRow >= this.data.length || prevRow < 0;
-      if (movingVertical && reachedVerticalBounds) return;
+      const previousActiveRow = Number(nextCell?.rowIndex || 0);
 
-      if (this.activeCellEditor) {
-        if (!this.activeCellCanClose()) return;
-        cellNode.endCellEdit();
-      }
+      if (key === 'ArrowUp') nextCell = nextCell?.cellAbove ?? null;
+      else if (key === 'ArrowDown') nextCell = nextCell?.cellBelow ?? null;
+      else if (key === 'ArrowLeft') nextCell = nextCell?.cellLeft ?? null;
+      else if (key === 'ArrowRight') nextCell = nextCell?.cellRight ?? null;
 
-      const activateCellNumber = cellNumber + cellDiff;
-      const activateRowIndex = rowDiff === 0 ? Number(this.activeCell?.row) : rowIndex;
-      this.setActiveCell(activateCellNumber, activateRowIndex);
+      const activeRowIndex = nextCell?.rowIndex || 0;
+      nextCell?.focus?.();
 
       // Handle row selection
+      const movingVertical = key === 'ArrowDown' || key === 'ArrowUp';
       if ((this.rowSelection === 'mixed' || this.rowSelection === 'multiple') && movingVertical && e.shiftKey) {
-        const previousActiveRow = Number(cellNode.parentElement.getAttribute('row-index'));
         this.#lastShiftedRow ??= previousActiveRow;
-        const shiftSelectFrom = Math.min(activateRowIndex, this.#lastShiftedRow);
-        const shiftSelectTo = Math.max(activateRowIndex, this.#lastShiftedRow);
+        const shiftSelectFrom = Math.min(activeRowIndex, this.#lastShiftedRow);
+        const shiftSelectTo = Math.max(activeRowIndex, this.#lastShiftedRow);
 
         if (Number.isNaN(shiftSelectFrom) || Number.isNaN(shiftSelectTo)) return;
 
@@ -805,14 +805,12 @@ export default class IdsDataGrid extends Base {
       }
 
       if (this.rowSelection === 'mixed' && this.rowNavigation) {
-        (cellNode.parentElement as IdsDataGridRow).toggleRowActivation();
+        (nextCell?.parentElement as IdsDataGridRow).toggleRowActivation();
       }
-      e.preventDefault();
-      e.stopPropagation();
     });
 
     // Handle Selection and Expand
-    this.listen([' '], this, (e: Event) => {
+    this.listen([' ', 'Space'], this, (e: Event) => {
       if (this.openMenu) return;
       if (this.activeCellEditor) return;
       if (!this.activeCell?.node) return;
@@ -842,33 +840,19 @@ export default class IdsDataGrid extends Base {
 
     // Follow links with keyboard and start editing
     this.listen(['Enter'], this, (e: KeyboardEvent) => {
-      if (this.openMenu) return;
-      if (!this.activeCell?.node || findInPath(eventPath(e), '.ids-data-grid-header-cell')) return;
-      if (!this.activeCellCanClose()) return;
+      // if (this.openMenu) return;
+      // if (!this.activeCellCanClose()) return false;
 
-      const row = this.rowByIndex(this.activeCell.row)!;
-      if (!row || row.disabled) return;
-
-      const cellNode = this.activeCell.node;
-      const hyperlink = cellNode.querySelector('ids-hyperlink');
-      const button = cellNode.querySelector('ids-button');
-      const customLink = cellNode.querySelector('a');
-
-      if (hyperlink && !hyperlink.container.matches(':focus') && !hyperlink.hasAttribute('disabled')) {
-        hyperlink.container.click();
-        hyperlink.container.focus();
+      const cellNode = this.cellLastActive;
+      if (!cellNode?.isEditing) {
+        cellNode?.startCellEdit();
+        return;
       }
 
-      if (button && !button.hasAttribute('disabled')) {
-        button.click();
-      }
+      const nextCell = e.shiftKey ? cellNode?.cellAbove : cellNode?.cellBelow;
 
-      customLink?.click();
-
-      if (customLink) {
-        cellNode.focus();
-      }
-      this.#handleEditMode(e, cellNode);
+      if (this.editNextOnEnterPress) nextCell?.startCellEdit();
+      if (!nextCell?.isEditing) nextCell?.focus?.(); // focus non-editable cells
     });
 
     // Commit Edit
@@ -893,18 +877,18 @@ export default class IdsDataGrid extends Base {
 
     // Edit Next
     this.listen(['Tab'], this, (e: KeyboardEvent) => {
-      if (this.openMenu) return;
-      if (this.activeCellEditor) {
-        if (!this.activeCellCanClose()) return false;
-        if (e.shiftKey) this.#editAdjacentCell(IdsDirection.Previous);
-        else this.#editAdjacentCell(IdsDirection.Next);
-
-        e.stopImmediatePropagation();
-        e.stopPropagation();
-        e.preventDefault();
-        return false;
+      // if (this.openMenu) return;
+      // if (!this.activeCellCanClose()) return false;
+      if (!this.isEditable) {
+        return;
       }
-      return true;
+
+      e.stopImmediatePropagation();
+      e.stopPropagation();
+      e.preventDefault();
+      const nextCell = e.shiftKey ? this.cellLastActive?.cellLeftEditable : this.cellLastActive?.cellRightEditable;
+      if (this.editNextOnEnterPress) nextCell?.startCellEdit?.();
+      if (!nextCell?.isEditing) nextCell?.focus?.(); // focus non-editable cells
     });
 
     this.offEvent('keydown.body', this.header);
@@ -921,81 +905,6 @@ export default class IdsDataGrid extends Base {
       }
     });
     return this;
-  }
-
-  /**
-   * Find the next editable cell and start editing it
-   * @private
-   * @param {IdsDirection} direction The cell element
-   * @returns {IdsDataGridCell} IdsDataGridCell
-   */
-  #editAdjacentCell(direction: IdsDirection): IdsDataGridCell {
-    this.commitCellEdit();
-
-    let nextCell = direction === IdsDirection.Next
-      ? next(this.activeCell.node, '.is-editable') as IdsDataGridCell
-      : previous(this.activeCell.node, '.is-editable') as IdsDataGridCell;
-
-    const rows = this.body?.querySelectorAll('.ids-data-grid-row');
-    if (!nextCell && rows && direction === IdsDirection.Next) {
-      for (let index = this.activeCell.row + 1; index < rows.length; index++) {
-        const row = rows[index];
-        if ((row.firstChild as Element).classList?.contains('is-editable')) {
-          nextCell = row.firstChild as IdsDataGridCell;
-          break;
-        }
-        nextCell = next(row.firstChild, '.is-editable') as IdsDataGridCell;
-        if (nextCell) break;
-      }
-    }
-
-    if (!nextCell && rows && direction === IdsDirection.Previous) {
-      for (let index = this.activeCell.row - 1; index >= 0; index--) {
-        const row = rows[index];
-        if ((row.lastChild as Element).classList?.contains('is-editable')) {
-          nextCell = row.lastChild as IdsDataGridCell;
-          break;
-        }
-        nextCell = previous(row.lastChild, '.is-editable') as IdsDataGridCell;
-        if (nextCell) break;
-      }
-    }
-
-    if (!nextCell) {
-      if (this.addNewAtEnd) {
-        this.addRow({});
-        return this.#editAdjacentCell(IdsDirection.Next);
-      }
-
-      this.activeCell.node.focus();
-      this.activeCell.node.startCellEdit();
-      return this.activeCell.node;
-    }
-
-    const row = Number(nextCell.parentElement?.getAttribute('aria-rowindex')) - 1;
-    const cell = Number(nextCell.getAttribute('aria-colindex')) - 1;
-    this.setActiveCell(cell, row, true);
-    nextCell.startCellEdit();
-    return nextCell;
-  }
-
-  /**
-   * Save or start editing
-   * @param {KeyboardEvent} e The cell event
-   * @param {IdsDataGridCell} cellNode The cell element
-   * @private
-   */
-  #handleEditMode(e: KeyboardEvent, cellNode: IdsDataGridCell) {
-    // Editing Keyboard
-    if (this.editable && cellNode.classList.contains('is-editable') && cellNode.classList.contains('is-editing')) {
-      cellNode.endCellEdit();
-      cellNode.focus();
-
-      if (this.editNextOnEnterPress) {
-        this.setActiveCell(Number(this.activeCell?.cell), Number(this.activeCell?.row) + (e.shiftKey ? -1 : 1));
-        this.activeCell.node.startCellEdit();
-      }
-    } else cellNode.startCellEdit();
   }
 
   /**
@@ -2436,11 +2345,15 @@ export default class IdsDataGrid extends Base {
     }
   }
 
+  get isEditable() {
+    return this.columns?.some((column) => !!column?.editor);
+  }
+
   /**
    * Edit the first editable cell
    */
   editFirstCell() {
-    this.#editAdjacentCell(IdsDirection.Next);
+    if (this.isEditable) this.cellLastActive?.cellRightEditable?.startCellEdit();
   }
 
   /**
@@ -3209,14 +3122,11 @@ export default class IdsDataGrid extends Base {
    * Checks on the active cell editor to see if its current state allows it to be closed.
    * @returns {boolean} true if the cell editor is able to "close"
    */
-  private activeCellCanClose() {
-    if (this.activeCellEditor && this.activeCellEditor.editor) {
-      if (this.activeCellEditor.editor.popup) {
-        if (this.activeCellEditor.editor.popup.visible) {
-          return false;
-        }
-      }
-    }
+  private activeCellCanClose(): boolean {
+    if (this.activeCellEditor?.isEditing) return false;
+    if (this.activeCellEditor?.editor?.popup?.visible) return false;
+    if (this.activeCellEditor?.column?.editor?.inline) return false;
+
     return true;
   }